+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Added help page.
+- [SQL Scheduler] Move "Running..." from "Last Execution" into "Next Due" column.
+- [SQL Scheduler] Fix "Invalid Date" in job logs.
 
 ## 2024-02-01 - 0.4.5